requirement.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
numpy==1.14.2
opencv-contrib-python==3.4.0.12
opencv-python==3.4.0.12
Pillow==5.0.0
pytesseract==0.2.0
requests==2.18.4
selenium==3.11.0
urllib3==1.22
